requirements.txt
requests>=2.31.0
pandas>=2.0.0
gspread>=6.0.0
google-auth>=2.25.0
python-dotenv>=1.0.0
